tools.py
import os
from dotenv import load_dotenv
load_dotenv()

from PyPDF2 import PdfReader

class BloodTestReportTool:
    @staticmethod
    async def read_data_tool(path='data/sample.pdf'):
        """Load and parse a blood test report PDF file."""
        text = ""
        reader = PdfReader(path)
        for page in reader.pages:
            text += page.extract_text() + "\n"
        return [{"page_content": text}]

class NutritionTool:
    @staticmethod
    async def analyze_nutrition_tool(blood_report_data):
        """Analyze nutrition based on blood test report data."""
        return "This is a placeholder nutrition analysis."

class ExerciseTool:
    @staticmethod
    async def create_exercise_plan_tool(blood_report_data):
        """Create an exercise plan based on blood test report data."""
        return "This is a placeholder exercise plan."
