requirements.txt
crewai==0.130.0
## crewai-tools==0.47.1
fastapi==0.110.3
python-dotenv==1.0.1
uvicorn==0.29.0
openai==1.68.2
PyPDF2==3.0.1